Minor fixes to class_table_inheritance_constraint_validations plugin
Raise error if attempting to load it into a model that doesn't
already load both the class_table_inheritance and constraint_validations
plugins.  In general, the preference is for plugins to load other plugins
they depend on, but in this case, both class_table_inheritance and
constraint_validations may need options, and passing the same option hash
to both plugins isn't a good idea.

Only call parse_constraint_validations_dataset if the model has a dataset.
This avoids both methods having to check for a dataset.

Fix handling of databases that do identifier mangling by default, such as
MSSQL.  Previously, the plugin was broken on such databases.

Avoid unnecessary local variable definition in parse_constraint_validations.

Update CHANGELOG, and have the plugins page link to the plugin.