Support num_waiting for the timed_queue and shared_timed_queue connection pools
This is useful to see how many threads/fibers are waiting to
checkout a connection. This can be useful when logging, so you can
see how the value changes over time. By checking this, you can see
if you could potentially benefit from a larger connection pool size.

This would be nice to support in the threaded connection pools, but
those use ConditionVariable instead of Queue, and Ruby doesn't
expose how many threads are waiting on a ConditionVariable.