Skip class_table_inheritance_constraint_validations specs when database is frozen
The individual specs were already skipped, but it still ran the
before(:all) code, which broke as it tries to change the
table used, which doesn't work if the database is frozen.

While here, use a shorter table name. The current table name is
too long to work on Oracle 11.