Support VIRTUAL generated columns on PostgreSQL 18+ using :virtual option
PostgreSQL 18+ made the default VIRTUAL if not specified. This doesn't
break backwards compatibility, as STORED was required previously.

Unfortunately, making VIRTUAL the default in Sequel would break
backwards compatibility. So in Sequel, the default will remain
to have STORED as the default, and require the :virtual option to
use VIRTUAL.