Do not auto parameterize PostgreSQL range and multirange types containing expressions
In general, you can only parameterize values, and ranges and
multiranges can contain expressions. Check whether the start and
end of the range, and whether the elements of the multirange, are
themselves parameterizable before using auto parameterization.

Note that even if auto parameterization is skipped, range
literalization will only work correctly for ranges with types.
If a range doesn't have a type, Sequel will try to literalize
the value embedded in a string, which won't work for expressions.
This doesn't appear to be fixable, because PostgreSQL doesn't
offer a function to create an untyped range.