Add pg_auto_parameterize_duplicate_query_detection extension for N+1 query detection
This can detect N+1 queries in a more general way, compared to the
forbid_lazy_load plugin (which only detects N+1 queries related
to associations). It's still not perfect, and can result in missed
N+1 query reports if the query SQL differs. The reason this has
pg_auto_parameterize in the name is that while it does not really
interact with pg_auto_parameterize, it checks for duplicate queries,
and N+1 query SQL outside of pg_auto_parameterize or prepared statements/
bound variables would generally vary per-call.