Fix Database#listen in postgres adapter to ignore all notifications after UNLISTEN
Previously, there was a race condition. If not all notifications
received between the LISTEN and UNLISTEN were handled, subsequent
notifications would stick around until the next time listen (or
wait_for_notify) was called, and returned at that point, which
could be for the previous channel and not the current channel.

Use a loop to ensure that all notifications received have been
ignored after the UNLISTEN, so the next LISTEN starts with a
clean slate.