Switch PostgreSQL Database#indexes :include_invalid option to :invalid
The :include_invalid support has been switched to invalid: :include.
This adds support for requesting only invalid indexes using
invalid: :only. I probably should have done something similar when
adding support for partial indexes. This doesn't change the
:include_partial option to :partial (that wouldn't be backwards
compatible), though I'm willing to add support for partial: :include
and partial: :only if anyone wants that ability.

This removes the :include_invalid spec instead of modifying it. The
spec only worked correctly if the user running the specs is the
database superuser. That may be true in CI, but it's not recommended
for standard testing during development. I'm not sure how to force
an invalid index in all supported PostgreSQL versions, and since
this updates the add_index :only option specs to test the indexes
:invalid option, I don't think the separate spec is needed.

Since it's simple, add a mock adpater spec for use of the add_index
:only option.