Disallow passing a set to Datset#where/exclude and other methods
Unlike arrays, Sequel does not treat sets as condition specifiers,
and passing a non-condition specifier array as an argument to
these methods results in an exception, so passing a set should
result in the same extension.

The set_literalizer extension added support for sets as condition
specifiers, so update it to make it backwards compatible.