Require sequel_pg/model when loading models if sequel_pg 1.18+ is loaded
This will automatically optimize the case where sequel/core is
loaded, then sequel/adapters/postgres (with sequel_pg), then
sequel/model. Previously in that case, the model code optimizations
would be skipped.

This is done only if sequel_pg 1.18+ is used, because that is the
first version that splits the model code into its own file.