Add single_statement_dataset_destroy plugin, so that dataset.destroy uses a single DELETE statement
Historically, Sequel has always used a separate DELETE query per
row, since dataset.destroy was the same as all(&:destroy). However,
the only readon to use destroy over delete is to run the destroy
hooks. If you don't have around_destroy hooks, you can run the
before_destroy hooks for all instances, then do a single DELETE
statement, then run the after_destroy hooks for all instances.

If the model defines a custom around_destroy hook, this falls
back to the default behavior.

This has a check that the number of rows returned by the all
matches the number of rows deleted. If not, an exception is
raised and the changes are rolled back. This will not catch
a case where an existing entry is removed and a new one is
added, though.